refactor(plugin-inbox): move list handlers into google/calendar and google/people
Move get-google-calendars → google/calendar/list
Move get-google-contact-groups → google/people/list

Co-locates the discovery/list handlers with the sync handlers in each
Google API subdirectory, consistent with the module structure.
