+++ packages/plugins/plugin-inbox/src/operations/google/people/index.ts
@@ -4,12 +4,18 @@
 
 import { type Operation, OperationHandlerSet } from '@dxos/compute';
 
+import List from './list';
 import Sync from './sync';
 
 export const ContactsFunctions: {
+  List: Operation.WithHandler<Operation.Definition.Any>;
   Sync: Operation.WithHandler<Operation.Definition.Any>;
 } = {
+  List,
   Sync,
 };
 
-export const ContactsHandlers = OperationHandlerSet.lazy(() => import('./sync'));
+export const ContactsHandlers = OperationHandlerSet.lazy(
+  () => import('./list'),
+  () => import('./sync'),
+);
