+++ packages/plugins/plugin-inbox/src/operations/index.ts
@@ -12,9 +12,9 @@ export const InboxOperationHandlerSet = OperationHandlerSet.lazy(
   () => import('./draft-email-and-open'),
   () => import('./draft-email'),
   () => import('./extract-contact'),
-  () => import('./get-google-calendars'),
-  () => import('./get-google-contact-groups'),
+  () => import('./google/calendar/list'),
   () => import('./google/calendar/sync'),
+  () => import('./google/people/list'),
   () => import('./google/people/sync'),
   () => import('./google/gmail/send'),
   () => import('./google/gmail/sync'),
